MNT: put test helper functions into their proper place
